% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium_bands.R
\name{visium_bands}
\alias{visium_bands}
\title{Creates bands inside and outside of tissue.}
\usage{
visium_bands(tpl)
}
\arguments{
\item{tpl}{The tissue position list table, a data.table}
}
\value{
tpl with "band" column, a data.table
}
\description{
Adds a "band" column to the tissue position list table.
}
\examples{
R code here showing how your function works
}
\keyword{spatial}
